Made a basic HVAC interface
Connor and I can take the basic interface and hook it into the respective Controller and Model to allow functionality to the interface now that I have created it. It looks dodgy since I tried to follow the general layout of what was done by David on the main_menu.fxml. I didn't create any images as I am not that artsy so it looks really rough.