FINALLY WORKING
ALright, after 8 hours of messing around with this, the main menu works to take you to the hvac menu and the back button will take you back. THe HVAC Controller and Model don't do anythign right now, and I believe you have to create a related activity in the manifest file to make them work properly as it's looking for the called methods in the xml files in the main activity (I think). VehicleInfotainmentSystem2 is the folder you want now, as the other crashes pretty much no matter what and after 6 hours I gave up and just rebuilt it in a fresh project.